Enable uploading of spreadsheet files
This patch introduces Special:ImportSpreadsheet, which enables the
uploading of spreadsheet files (Excel, OOCalc, Gnumeric, ...) if the
PHPExcel package is installed.

This can be done using composer.  From the MediaWiki installation
directory call
      composer require "phpoffice/phpexcel:~1.8"

If PHPExcel is not installed, the special page will just not be available.

Minor enhancement: Columns with an empty header will be ignored instead
of resulting in an error message. This avoids cases where Excel saves
additional empty columns.

Change-Id: I65d9c3ba87a1a44665926a9acc35d12955822a89